+++ .takumi/plugin.js
@@ -1,0 +1,14 @@
+
+export default {
+  commands: async () => {
+    return {
+      name: 'foo',
+      description: 'foo',
+      fn: async () => {
+        console.log('foo');
+      },
+    };
+  },
+  cliStart: async () => {
+  },
+};